{"schema":"decodex/commit/1","summary":"Modularize settings capture frame controls","authority":"manual"}
